Add initial support for SystemD Unit integration
Keeping spawnpoint containers running depends on the spawnd process
being alive and well to maintain that state machine. Upon a peer
disconnect, spawnd can sometimes take a long time to reboot (and appears
to sometimes time-out, especially when the server in question is running
on spinning metal drives, which can cause issues for querying block
chain state in order to boot spawnd).

As per the design of spawnpoint, the *running* of containers is
decoupled from spawnd being up, but in practice the *restarting* of
containers is coupled with spawnd. To try and work around this, I've
hooked into spawnpoints container starting code to also generate a
simple SystemD unit file which should keep the spawnpoint container up
and running even if spawnd has issues starting up.

Eventually, I will make using systemd configurable so that spawnd
doesn't have to rely upon it, but for now I want to try and explore this
avenue in order to increase the reliability of spawnpoint.